+++ src/test/service/bounty/create/validate.test.ts
@@ -1,0 +1,5 @@
+describe('BountyCreateTest', () => {
+	describe('todo', () => {
+		console.log('todo');
+	});
+});