+++ src/app/events/guildMemberUpdate.ts
@@ -18,8 +18,10 @@ module.exports = {
 			const newMemberHasGuestPass = newMember.roles.cache.some(role => role.name === constants.DISCORD_ROLE_GUEST_PASS);
 			const oldMemberHasGuestPass = oldMember.roles.cache.some(role => role.name === constants.DISCORD_ROLE_GUEST_PASS);
 			const isNowActive: boolean = (newMemberHasGuestPass && !oldMemberHasGuestPass);
-			console.log(`guest pass active for username ${newMember.user.username}: ${isNowActive}`);
-			updateNotionGuestPassDatabase(newMember.user.tag, isNowActive);
+			if (isNowActive) {
+				console.log(`guest pass active for username ${newMember.user.username}`);
+			}
+			return updateNotionGuestPassDatabase(newMember.user.tag, isNowActive);
 		}
 	},
 };