+++ src/app/utils/db.ts
@@ -9,19 +9,17 @@ const state: {db: Db, client: MongoClient, mode} = {
 
 const db = {
 	connect(database: string, done: (error?: MongoError) => Promise<any>): Promise<any> {
-		console.log('connecting to DB');
 		try {
 			MongoClient.connect(
 				constants.MONGODB_URI_PARTIAL + database + constants.MONGODB_OPTIONS,
 				{ useUnifiedTopology: true },
 				async (err: MongoError, client: MongoClient) => {
 					if (err) {
-						return await done(err);
+						return done(err);
 					} else {
-						console.log('connected to DB');
 						state.db = client.db(database);
 						state.client = client;
-						return await done();
+						return done();
 					}
 				},
 			);
@@ -35,7 +33,6 @@ const db = {
 	},
 
 	close(): Promise<void> {
-		console.log('closing connection to DB');
 		return state.client.close();
 	},
 };