+++ src/test/service/bounty/claim.test.ts
@@ -1,15 +1,15 @@
 import * as chai from 'chai';
 import claim from '../../../app/service/bounty/claim';
 import * as sinon from 'sinon';
+import ServiceUtils from '../../../app/utils/ServiceUtils';
 import { MongoClient } from 'mongodb';
-import serviceUtils from '../../../app/service/ServiceUtils';
 import { DiscordAPIError } from 'discord.js';
 
 const assert = chai.assert;
 
-describe('BountyClaimTest', () => {
+describe('BountyClaim', () => {
 	let ctx;
-	let serviceUtilsStub;
+	let serviceUtilsMock;
 
 	beforeEach(() => {
 		ctx = {
@@ -25,8 +25,11 @@ describe('BountyClaimTest', () => {
 			send: (message: string) => { return message; },
 		};
 
-		serviceUtilsStub = sinon.stub(serviceUtils, 'getGuildMember');
-		serviceUtilsStub.returns({ send: (message) => { return message; } });
+		serviceUtilsMock = sinon.mock(ServiceUtils);
+		serviceUtilsMock.expects('getGuildAndMember').returns({
+			guild: {},
+			guildMember: { send: (message) => { return message; } },
+		});
 	});
 
 	afterEach(() => {
@@ -37,29 +40,30 @@ describe('BountyClaimTest', () => {
 
 		it('should be invalid bountyId for null', async function() {
 			ctx.options.claim['bounty-id'] = null;
-			const result = await claim(ctx);
-			assert.equal(result, '<@567865362541182987>\n' +
-			'Please enter a valid bounty hash ID: \n' +
-			' - can be found on bountyboard website\n' +
-			' - https://bankless.community');
+
+			try {
+				await claim(ctx);
+			} catch (e) {
+				assert.equal(e.message, 'invalid bountyId');
+			}
 		});
 
 		it('should be invalid bountyId full special character', async function() {
 			ctx.options.claim['bounty-id'] = '$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$';
-			const result = await claim(ctx);
-			assert.equal(result, '<@567865362541182987>\n' +
-				'Please enter a valid bounty hash ID: \n' +
-				' - can be found on bountyboard website\n' +
-				' - https://bankless.community');
+			try {
+				await claim(ctx);
+			} catch (e) {
+				assert.equal(e.message, 'invalid bountyId');
+			}
 		});
 
 		it('should be invalid bountyId full negative numbers', async function() {
 			ctx.options.claim['bounty-id'] = '-10005';
-			const result = await claim(ctx);
-			assert.equal(result, '<@567865362541182987>\n' +
-				'Please enter a valid bounty hash ID: \n' +
-				' - can be found on bountyboard website\n' +
-				' - https://bankless.community');
+			try {
+				await claim(ctx);
+			} catch (e) {
+				assert.equal(e.message, 'invalid bountyId');
+			}
 		});
 
 	});
@@ -68,18 +72,19 @@ describe('BountyClaimTest', () => {
 		it('should be mongodb error', async () => {
 
 			const mock = sinon.mock(MongoClient);
-			mock.expects('connect').once().throws('bad connection');
+			mock.expects('connect').throws('bad connection');
 
-			const result = await (await claim(ctx));
-
-			assert.equal(result, 'Sorry something is not working, our devs are looking into it.');
+			try {
+				await claim(ctx);
+			} catch (e) {
+				assert.equal(e.message, 'Sorry something is not working, our devs are looking into it.');
+			}
 		});
 
 		it('should be client api error', async () => {
-			serviceUtilsStub.restore();
-
-			const mock = sinon.mock(serviceUtils);
-			mock.expects('getGuildMember').once().throws(new DiscordAPIError('', new Error('Mock Discord API Error'), 'GET', 405));
+			serviceUtilsMock.restore();
+			const mock = sinon.mock(ServiceUtils);
+			mock.expects('getGuildAndMember').once().throws(new DiscordAPIError('', new Error('Mock Discord API Error'), 'GET', 405));
 
 			try {
 				await (await claim(ctx));