+++ src/test/service/bounty/create/validate.test.ts
@@ -1,5 +1,0 @@
-describe('BountyCreateTest', () => {
-	describe('todo', () => {
-		console.log('todo');
-	});
-});