+++ src/test/service/bounty/list.test.ts
@@ -1,11 +1,13 @@
 import * as chai from 'chai';
 import * as sinon from 'sinon';
 import list from '../../../app/service/bounty/list';
+import ServiceUtils from '../../../app/utils/ServiceUtils';
 
 const assert = chai.assert;
 
-describe('BountyListTest', () => {
+describe('BountyList', () => {
 	let ctx;
+	let serviceUtilsMock;
 
 	beforeEach(() => {
 		ctx = {
@@ -21,19 +23,25 @@ describe('BountyListTest', () => {
 			send: (message: string) => { return message; },
 		};
 
-		afterEach(() => {
-			sinon.restore();
+		serviceUtilsMock = sinon.mock(ServiceUtils);
+		serviceUtilsMock.expects('getGuildAndMember').returns({
+			guild: {},
+			guildMember: { send: (message) => { return message; } },
 		});
+	});
+
+	afterEach(() => {
+		sinon.restore();
 	});
 
 	describe('Parameter Validation', () => {
 
 		it('should be invalid bounty-type', async function() {
 			ctx.options.list['list-type'] = 'sadfasdfsdaf';
+
 			try {
 				await list(ctx);
 			} catch (e) {
-				console.log(e);
 				assert.equal(e.message, 'invalid bounty type');
 			}
 		});