+++ src/app/commands/admin/GuestPass.ts
@@ -51,21 +51,20 @@ module.exports = class GuestPass extends SlashCommand {
 		const guildMember = await guild.members.fetch(ctx.options.user);
 
 		if (guildMember.user.bot) {
-			ctx.send('Bots don\'t need a guest pass!');
-			return;
+			return ctx.send('Bots don\'t need a guest pass!');
 		}
 
 		// Retrieve Guest Pass Role
-		const guestRole = await retrieveGuestRole(guild.roles);
+		const guestRole = retrieveGuestRole(guild.roles);
 
 		// Open database connection
-		db.connect(constants.DB_NAME_DEGEN, async (err) => {
+		await db.connect(constants.DB_NAME_DEGEN, async (err) => {
 			if (err) {
 				console.error('ERROR:', err);
 				return;
 			}
 			// DB Connected
-			const dbGuestUsers = await db.get().collection(constants.DB_COLLECTION_GUEST_USERS);
+			const dbGuestUsers = db.get().collection(constants.DB_COLLECTION_GUEST_USERS);
 			const queryOptions = {
 				upsert: true,
 			};
@@ -91,7 +90,7 @@ module.exports = class GuestPass extends SlashCommand {
 			// Add role to member
 			guildMember.roles.add(guestRole).catch(console.error);
 			console.log(`user ${guildMember.id} given ${constants.DISCORD_ROLE_GUEST_PASS} role`);
-			ctx.send(`Hey <@${guildMember.id}>! You now have access for ${expiresInHours / 24} days.`);
+			return ctx.send(`Hey <@${guildMember.id}>! You now have access for ${expiresInHours / 24} days.`);
 		});
 
 		// Send out notification on timer