+++ src/app/events/guildMemberUpdate.ts
@@ -19,7 +19,10 @@ module.exports = {
 			const oldMemberHasGuestPass = oldMember.roles.cache.some(role => role.name === constants.DISCORD_ROLE_GUEST_PASS);
 			const isNowActive: boolean = (newMemberHasGuestPass && !oldMemberHasGuestPass);
 			console.log(`guest pass active for username ${newMember.user.username}: ${isNowActive}`);
-			updateNotionGuestPassDatabase(newMember.user.tag, isNowActive);
+			const updateNotionPromise = updateNotionGuestPassDatabase(newMember.user.tag, isNowActive);
+			updateNotionPromise.catch(error => {
+				console.log(error);
+			});
 		}
 	},
 };