+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.1-RELEASE (2021-07-28)
+
+1. Handle all notion api calls in case of rate limit or out of sync errors
+
 ## 1.1.0-RELEASE (2021-07-21)
 
 1. Add typescript integration and reorder events initialization